+++ src/ambition/version_details.cc
@@ -18,6 +18,12 @@ std::vector<VersionUpdates> versionDetails = {
     .identifier = _(IDENTIFIER_PREVIEW),
     .sections = {
       {
+        .heading = _(HEADING_ADDED),
+        .items = {
+          _("Added kingdom-to-kingdom score comparison to the Rank Report, allowing you to see the scores of two kingdoms side-by-side."),
+        },
+      },
+      {
         .heading = _(HEADING_CHANGED),
         .items = {
           _("Changed default keybind for train soldier to `F` from `W`, to make it more consistent with the other fort-related keybinds (build fort)."),