+++ CHANGELOG.md
@@ -10,6 +10,11 @@ speaking, considered to be the player's experience.
 
 ## [Unreleased]
 
+### Added
+
+- Added kingdom-to-kingdom score comparison to the Rank Report, allowing you to
+  see the scores of two kingdoms side-by-side.
+
 ### Changed
 
 - Changed default keybind for train soldier to `F` from `W`, to make it more